requirements.txt
cycler==0.10.0
dnspython==1.16.0
enum-compat==0.0.2
eventlet==0.20.1
freetype-py==2.1.0.post1
greenlet==0.4.15
h5py==2.10.0
kiwisolver==1.0.1
mysql-connector-python==8.0.19
numpy==1.18.2
packaging==20.3
protobuf==3.6.1
PyOpenGL==3.1.0
pyparsing==2.2.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyqtgraph==0.10.0
pyserial==3.4
python-dateutil==2.6.0
sip==5.1.1
six==1.10.0
toml==0.10.0
vispy==0.6.4
